kern.sire: Syscalls use CNSTR naming convention.
Syscalls previously used the TypesNamingConvention and we previously
defined a special function per syscall.

Now that we are building types for real, it's worthwhile to change this
so that we can maintain consistent naming conventions throughout.

The old approach looked like:

    \ TimeWhen : IO TimeStamp

    = (ex pure)
    : now <- TimeWhen
    : ()  <- TimeWait (add now 1)
    | pure ()

Instead, we treat each syscall binding as a constructor for a `SysCall`
datatype, that you can run using the `syscall` function.

    \ TimeWhen : SysCall TimeStamp

    \ TimeWait : TimeStamp -> SysCall Unit

    \ syscall  : SysCall a -> IO a

    = (ex pure)
    : now <- syscall TIME_WHEN
    : ()  <- syscall (*TIME_WAIT (add now 1))
    | pure ()