boot.sire: col macro: better named-continuation syntax
With the previous code, the LHS of a `:` binding would be a pinned,
named laws, if the first words of the signature was capitalized.

    : (Foo x y) <- expr

This looks nice, but is totally inconsistent with the other naming
conventions, and it's a very unintuitive edge-case.

This commit replaces that syntax with

    : ?!(foo x y) <- expr

This, while uglier, directly matches the underlying lambda syntax:

    : ?!(foo x y) <- expr
    | foo

        ==>

     | exp
    ?! (foo x y)
     | foo

For consistency, the other two lambda runes are accepted as well:

    : ?(foo x y) <- expr

    : &(foo x y) <- expr