sire: Use `*` for macro branches.
Previously, we have been using `=` for switch branches and similar
syntactic forms.

    = foo 3
    = bar 4
    = (bar x)
    # switch x
    = 0 'Zero'
    = 1 'One'
    = 2 'Two'
    'Other'

This is pretty, but I find it somewhat jarring that the branch rune is
the sames as the top-level-definition rune.

This commit swaps that out for `*`:

    = foo 3
    = bar 4
    = (bar x)
    # switch x
    * 0 'Zero'
    * 1 'One'
    * 2 'Two'
    'Other'

This seems about as aesthetic as `=` while also being clearly
distinguished from the top-level bindings.

`*` is already in use for inlining.  However, like with branches, the
inlining-case of `*` is a "decoration" for another syntactic form (The
`|` rune).  Also, inlining and branching are very different (visually
and conceptually), so I feel that this isn't confusing.