sire: datacase: Backfilled support for Either, Bit
Either and Bit are defined before `#datatype` and `#datacase`, but we
can manually set the syntax-properties that `#datacase` needs.

I can't yet support lists, maybe, etc because those datatypes are not
tagged, and the current `#datatype` code expects something tagged.
I'm working on generalizing that to work with lengths/etc.