sire: nicer tab syntax
This makes the following syntax into a tab literal:

    [x=3]

The above is treated as a short-hand for:

    %[x=3]

This is unambiguous because `=` is never used as an expression or
pattern rune.  It is only used for commands and as decoration syntax
for things like tabs.

This is especially nice with prefix or tight-infix notation:

    | trk %func,[=x =y =z]

        ==>

    | trk %func,(%[=x =y =z])

And

    = (foo x=[=a =b =c])

        ==>

    = (foo x=(%[=a =b =c]))