sire: !!= blocks must have only !!= commands
This change allows the heir of !!= to be used as a second son:

    !!= (add 3 4)
      | add 4 3

Previously, you could mix !!= commands with other commands in the
same block:

      = x 3
    !!= x x
      = y x

This is ugly, and it makes the !!= parsing code more difficult.

In particular, because the heir could be a macro (that expands to a
command), there is no way to determine if the heir is an extra argument,
or if the heir is another command.

I considered eliminating multiple-commands-per-block altogether, but
both of these options seem very ugly to me:

    !!= 1 | inc ( add 0   0   )

    !!= 1 | inc ( add 0   0-0 )

    !!= 1 | inc ( add 0-0 0   )

    !!= 1 | inc ( add 0-0 0-0 )

    !!= 1 | add 0 1

    !!= 1 | add 1 0

And

    * !!= 1 | inc ( add 0   0   )
    * !!= 1 | inc ( add 0   0-0 )
    * !!= 1 | inc ( add 0-0 0   )
    * !!= 1 | inc ( add 0-0 0-0 )
    * !!= 1 | add 0 1
    * !!= 1 | add 1 0

So, I decided to keep multiple-commands-per-block, but only in the
limited case where every command is of the same type (same rune).