sire: * as multi-command rune
This introduces the `*` rune as a multi-command rune.  For example:

    * x=3
    * y=3
    * !!= x y
    * = (f x y)
      | x
    * (f x)

This is mostly useful for command-macros which want to expands to more
than one command.

    = [x y] 3,4

        ==>

    * #= _tmp 3,4
    * #= x (idx 0 _tmp)
    * #= y (idx 1 _tmp)

Another use-case is macros that want to have an effect, but don't actually
want to expand to anything.  For example,

    printType#add

        ==> prints: (Nat -> Nat -> Nat)

    0

But that causes `0` to be printed and `_` to be rebound to 0.  After this
change, it can just expand to `(*)`, which does nothing.