sire: switch .{quoted symbol} rune to dot
This moves the "quoted symbol" rune from '##' to '.' (or '#.').

This is in anticipation of future support for module-qualified references
(module.symbol):

    = cenMacro .{%}

    /+ json [printJson]

    | printJson
    | json.ROW
    | base.map json.not
    | [json.TRUE json.FALSE]

------------------------------------------------------------------------

This also changes the law-where-tag-is-different-from-symbol rune to "@".

Previously:

    (f##g x)=x => (g x)=x

Now:

    (f@g x)=x => (g x)=x

This doesn't conflict with the `@` syntax for let bindings nor for the
`@` syntax for patterns (because the function-name syntax is separate).

This is not beautiful, but it doesn't matter because it's an extremely
obscure edge-case.