sire: explicit file-load order
This introduces the `###` rune, which loads another sire file and
specifies the current modules name.

    ### 03_nat <- 02_bit

The above will cause 01_fan and 02_bit to load.  Each module may load
only zero or one other modules.

Also, this changes the behavior of the `/+` rune to not load files at all,
and to instead simply import symbols from already-loaded files.

    ### foo <- 02_bit

    /+ 01_fan [add] ; good

    /+ 04_app [len] ; error: not loaded

The purpose of this change is to force a linear load order, so that:

-   Everything is deterministic.  Previously, you would get different
    binding keys for a binding depending on where it was imported from
    (on which other files were loaded first).

-   We can cache Sire loading, instead of loading everything every time.

-   We can combine a bunch of sire files into a single input by just
    concatenating them.

Short term, this is somewhat annoying, since it means that we need to
explicitly linearise all of the tests and demos.  However, once caching
is working, we can just have all of those files be roots, and have them
all just depend on `prelude`.