sire: merge ECNS/ENAT, never let-bind constants
This eliminates the `ENAT` branches of the `Sire.Syntax.Exp` datatype
and just treats nat literals like any other PLAN constant.

This also changes the optimiation rules.  Previously:

    @ x "asdfkasjdfaoijaowiejfwef"
    x,x

Was not optimied away, since `x` is a large constant.  Also, any constants
*besides* number literals were also not optimized.

Now, the above let binding is optimized to:

    "asdfkasjdfaoijaowiejfwef","asdfkasjdfaoijaowiejfwef"

This is not wasteful because they will both just be pointers to the same
values (during complilation and also when loading pins from disk).

The main motivation for this change is to simplify the representation
of Exp and to simplify the rules around optimization.  Both in service
for freezing the thing.