sire: compiler uses 0-arity law intead of hack
Previously, the compiler compiled every input into a one-arity law and
then passed that a dummy argument.

    @ x 3
    @ y 4
    | add x y

        ==>

    ^ _ 0
    & x
    @ x 3
    @ y 4
    | add x y

It turns out that this was a workaround for a bug in the Loot backend.
The Loot bug was in some code that was itself doing *THE SAME* work-around
to deal with an ancient runtime system that couldn't handle zero-arity
laws.

Now, the inputs are evaluated by creating a zero-arity law, which just
evaluates immediatly with no dummy argument.

    @ x 3
    @ y 4
    | add x y

        ==>

    & (|)
    @ x 3
    @ y 4
    | add x y