datom: rewrite pull in cps style
During a discussion, Sol said this section would be much easier if I
rewrote it in with continuation passing. So I did. It's still a bit
complicated, though it is more than marginally easier to follow.

As part of this, changed the stateful foldl into a function which
just generated an error or an item to go into a tab instead of something
which mutated a list of errors and the resulting tab.