sire: tweaked inlining, ** rune, flexible arities
This makes a couple of simplifications to the inling system:

Previous, an expression of the form:

    (*f x y)

Would be inlined if f was a function that takes two arguments.

This was a little bit weird syntactically, and required some annoying
book-keeping to implement.

Now, an expression of the form `**f` means: "if this is a function that
is applied to enough arguments, inline it".

The uglier rune `**` is chosen because the more flexible syntax now
conflicts with the decorative uses of `*` (as a "bullet-points" syntax).

Another advantage is that the AST gets simpler, which is important
because I'm working towards making the AST part of the spec (in order
to give macros access to the inliner expression).

Differences in behavior:

1)  Arity checking is more flexible.  For example `(**K x y z)` is still
    inlined, even though it has too many arguments (two are used for
    inlining).

2)  Inline requests can be put anywhere, for example, in `(**I **K x
    y)`, the K is not an application head, but can still be marked as
    inline (and, in this situation, will inlined).

        (**I **K x y)
        (**K x y)
        x