sire: after name resolution EREC has selfreference
The difference between EREC and ELET is syntactic.  With `@@` the
self-reference is in scope, but with `@` it is not.

However, after name resolution, there was no meaningful difference except
that both branches needed to be handled by every pass.

Instead of creating a new representation that combines the two, this
changes makes the distinction meaningful.  The name resolution pass only
produces `EREC` if the binding is *actually* self-referential.

This matters in some cases, so it's useful.