stroll: initial commit
These are experiments in trying to build a bullet based outliner with
backlinks and other features found in Roam and logseq, with the hope of
driving datom development.

This starts with a few basic lookup pulls (some which really should be
:find queries, if we had that implemented), to try to drive thinking
about how all of this should work on the backend.