sire: rewrote compiler
This is a rewrite of the Sire compiler.  This significantly improves the
readability of the compiler code and fixes a ton of edge-cases.  There is
"morally" no change, however.  We still do the same set of optimizations.

A lot of details about the underlying PLAN code has changed, so this
breaks a bunch of jet hashes.

Some known changes in the output:

-   After inlining, lets were not optimized again.  For example, with
    `#mutrec` the helper function for each arm was always in the code,
    even if it was unused (or inlined into all of the call sites).

-   The order of lambda-lifted arguments is now directly tied to
    reference-order (given a left-first traversal of the function body).
    Previously, this was ordered based on the internal gensym key used
    by the thing.

-   Always-inline functions `(**I x ? x)` work if even not global,
    and work even if bound with a local let binding.

-   The previous code was too agressive (by mistake) about eliminiating
    let bindings, and sometimes ininlined bindings that were used twice.
    This resulted  in duplicated  work in some cases.