Fix licensing to be consistent on BSD-3.
Everything we've said in public is that we are permissively licensed,
but the only license reference in the repository was a single mention of
the AGPL in a file automatically generated from a template. Add the text
of BSD-3 as LICENSE and also adds a copyright notice to each haskell and
sire source file.