multicog: multiple Cogs in a Machine that can communicate
This is the first patch in a series that moves from single Cogs in an
event log to having multiple Cogs in a "Machine" which can communicate
over a hardware based ipc mechanism. If Cogs were to communicate over
the normal network interfaces, those network packets would have to be
written in whole into the event log, while if multiple cogs are part of
the event log, they can instead write very small backreferences like how
EVAL currently works.

The previous Actor based prototypes had actors share an event log, but
the reason that implementation was so complex was because of the actor
model meant that all outbox state was shared between actors, along with
the complexity of timeouts, cancellation and coherent snapshots. None of
this should be a problem for the Cog design.

This patch gets the implementation to the point where the database has
snapshots and log events associated with a "CogId" and the Cog
implementation keeps track of the state for each Cog.

No spawning/forking behaviour has been added in this patch, there's only
one cog with the identity `removeMeDummyCogId`.

Major todos:

- There's no forking or ipc implemented, this just changes storage and
  runtime interfaces.

- Hardware interface changes are hacked together, they need to be
  responsive to Cog spawning but aren't.

- Clean up the cli in ConsoleExe.hs. We need both different commands and
  to fix up things like the old du implementation.