lmdb: fix first batch number when there are no batches
In the ancient repository, this line started out as "newTVarIO (num + 1)"
for reasons lost to time. For reasons undocumented, this was changed to
the case statement that special cased "0" batches to 1, presumably due to
confusion about whether vNum was the number of batches or the next batch
number? But when there are 0 batches, there are not 1 batches.

This fixes crashes during the first logbatch writing when the daemon was
restarted before the first batch was written (ie, the cog was never spun).