plan: hack around off-by-one-error
Running all of the sire code with this check enabled but the corresponding
hack disabled will result in errors.

This should be resolved for real, but this change makes the problem go
away at the cost of one extra word per PLAN stack frame.

This adds a check to every PLAN call, but it doesn't seem to affect too
much, and it's just an int comparison.