multicog: remove first attempt at parallelism
This was trying to run multiple cogs at the same time concurrently,
but that runs into issues where we have to access two cogs at the
same time: in stop and in send/recv. While we could make a complicated
thing right now that blocks multiple threads, we should build the simple
thing first.