sire: sire.sire: remove EWUT
This removes the EWUT constructor from the `Exp` type in `sire.sire`.

The previous type-inference strategy was to partially parse the input
so that we can do type-checking on the parts that we understand.

However, the new strategy is to just fully parse all inputs, so the
"we don't know how to parse this" constructor no longer makes sense.