sire: 05_cmp: remove the length check from cmp
The `cmp` function previously compared the closures lengths in order to
make data-jet comparisons faster.

However, on further consideration, the check actually has no impact on
the behavior of the function.

This change also includes a comment-section documenting the
lazy-evaluation behavior of `cmp`.  Being precise about this is very
important w.r.t the correct implementation of a `cmp` jet.