sire: cab/tab segregation, syntax edition
This switches tab literals over to the # rune and ## runes, so you would
now write:

    #[] #[a=3] #[=a]

    ## =a 3
    ## =b 3

This removes the overlap betwen the tab/cab syntax which:

1)  Makes it possible to corrrectly type the empty tab/cab

2)  Makes it possible to fully separate cab/tabs in the future (so that
    the empty cab is not also a tab).

This also opens up the possibility of slightly more convient syntax,
now that there is no tension with the cab syntax.  For example, we can
drop the requirement for there to be an `=` on every key:

    #[a b=4]

    ## foo | f bar
    ## zaz | a b c
    ## lol | 1,2,3
    ## wut

This also reworks the cab and tab parsers to be much less of a tangle,
making it much more feasible to re-use that code from `stew.sire`.