multicog: random cog ids and cli cleanup.
This removes the hardcoded special 0th id, now all cog ids are randomly
generated.

Cleans up the CLI interface so that it speaks in terms of machines (and
cogs in a machine) instead of just cogs.

Fixes a few other bugs to get the fulltag demo working again with the above.