sire: untangle quasiquotation
The previous quasi-quotation macros was a massive tangle that I barely got working
and never could fully load into my head.

This commit manages to split the routine into a parser and a
code-generator, both with fairly straightforward code.

Also, this fixes a bug where quotations that appeared in the heir of a
leaf were ignored.  For example:

    foo($bar)

Finally, this new approach makes supporting quasi-quotation in Stew
possible.