stew: parse : rune
The previous : macro was hacky and didn't actually parse anything,
instead macro-expanding into something that would then be parsed by
other macros.

This builds a proper parser for `:`.  An early macro without
destructuring, and a later version with destructuring.  And finally,
stew re-uses that all of that to actually parse the thing.