stew: use more regular output in stew printer
This makes the stew printer output more regular (and less fancy).
The output is uglier in some cases, but this is intended to be a simple
debugging tool, not a pretty-printer.

The next commit in the patch-set will take advantage of the more-regular
output to improve the clarity of the tests.