blake3: don't build the main.c file as part of the library.
The blake3 main.c claims to only be runnable when its built with the
`make test` environment in the upstream build. Both it and the library
code use a preprocessor directive to control the visibility of internal
library symbols that the test binary then changes for testing.

This setup just breaks hard with how ghci does dynamic loading and the
easiest way to fix this is to just not including unused testing code in
a non-testing binary.