stew: fully hooked up expression compiler
This fully implements a Stew compiler for expressions, connects it the the
macro system, and switches all "downstream" code over to Stew.

There were a lot of tricky edge-cases that needed to be solved to get
everything working.

This doesn't yet do commands, only expressions, so things like the
following don't work:

    = (foo STATE(..)))
    ...

You must instead write:

    = (foo st)
    @ st STATE(..)
    ...

This is because the Stew compiler "kicks in" at the `@` rune.

This will be resolved soon, when Stew also learns how to handle commands.