cog: add %what to determine available hardware.
This adds %what to the list of valid requests in the Cog
interface. There was previously no way to detect what hardware is
available.

This proposes that we don't write *any* hardware into the base Cog spec,
that the Cog interface is just about how a function can do requests to a
runtime for parallelism, multiple processes and hardware access and
allows us to freeze the Cog interface soon once we resolve questions
about what `[%cog ...]` control commands should be included.

Under this proposal, the details of any underlying hardware become its
own standard for freezing later on as a separate layer of the stack: it
becomes a stack of PLAN, SIRE, COG, HARD. Just as people should be able
to take PLAN without taking the rest of the plunder stack, or PLAN+SIRE
without taking our machine interface, people should be able to take
PLAN+SIRE+COG without taking our proposed standardized hardware.