blake3: use gen in place of mkRow, delete `mkRow`
This removes the `mkRow` operation and reworks it's use-sites to use
`gen`.  `mkRow` is complicated, not type safe, and cannot be jetted.

This code was tested with the blake3 jet disabled.  I don't have a good
understanding of the blake3 functions I updated here, but I believe the
patch produces mechanically equivalent code.

This patch also includes some minor unrelated docs changes.