sire: add many type declarations an annotations
This imports the `types` module into these modules:

    demo_full_tag_site, datom, json, kern, hitch

And adds a bunch of type annotations and declarations to those modules.

None of this is doing type inference yet, but the type-names and
type-arities are checked.  This is part of a partial-merge of the
typed-sire-scratch code, pulling in bits that have stabilized to reduce
the distance between branches.