json: native representation
This path switches `json.sire` over to a "tagless" representation.
It uses the implicit tagging that comes from data-jets to determine the
JSON type (a bar is a string, a tab is a map, etc).

Support for `#datacase` on the representation is included, and some
call-sites are switched over to that usage-pattern (instead of explicitly
unpacking and switching on the tag).

This approach will be much slower until `_TypeTag` is jetted, that change
comes later in the patchset.

This chooses to represent numbers with (0 x) and to represent %null,
%true, and %false directly as nats.  Another alternative would have been
to represent numbers as nats and wrap null/true/false in COW or CAB,
but this approach seemed simplest.  Especially given that, once once
support is added for negative numbers and rationals, numbers will no
longer fit within the NAT type.