sire: effect-macros expand to #*
Previously "effect-only" macros like `#setProp` and `#printType` would
macro-exapand to "0".  This caused the REPL to print out a big string
of zeros whenever these commands were used.

This patch updates those macros to expand to `(#*)`, which is an "empty
sequence of commands".  The new form does not cause any REPL output.