multicog: keep state about spinning/crashed/timeout
This replaces the exponential backoff of trying to run a cog again with
`(NAT 0) %% val` with keeping track of whether a cog crashed/timed
out/etc as part of its formal state.

This `CogState` is passed as part of the value back to `[%cog %stop]` to
allow a cog to detect when a cog went bad.