plan: new cab/tab representation
This switches tabs and cabs over to a new legal representation.

This is a very tricky change, and likely contains at least one regression.
However, all of the tests pass.  I haven't tested the full-tag demo yet.

Previously, tabs and cabs were represented like:

    %[]        = (0 0 1 [])
    %[a]       = (0 0 2 [%a])
    %[a b]     = (0 0 3 [%a %b])
    #[]        = %[]
    #[a=1]     = (%[a] 1)
    #[a=1 b=2] = (%[a b] 2 1)

Now, they are represented as

    %[]        = (0 1 2 [])
    %[a]       = (0 1 2 [%a])
    %[a b]     = (0 1 2 [%a %b])
    #[]        = (%[] [])
    #[a=1]     = (%[a] [1])
    #[a=1 b=2] = (%[a b] [1 2])

This has a number of advantages:

-   Generally less tricky.

-   Much more efficient representation unjetted.

-   No overlap between empty tabs and empty cabs (a very annoying
    edge-case, since the interpreter must be very careful to never create
    an empty cab).