sire: rex-id: implement explicit binder keys
This patch brings sire one step closer to matching the spec.

Previously, binding keys were controlled by the rex-identity-key of the
symbol used to define the binding name.  This is confusing, and the rex
identity concept is slated for removal.

This change adds syntax to explicitly specify the desired binder key.

    #= 35 x 3           ; Binds x=3 with binder-key 35

    #= 38 (f x) x       ; Binds f=(f x ?? x) with binder-key 38