plan: simplify rex data-jet representation
The Rex data-jet was designed around the assumption that it had to
work as the first module in the bootstrapping system.  As a result, the
representation was complex, inefficient, and difficult to data-jet-match.

This is no longer necessary since the the need to put this so early
was resolved.  It was resolved by simply allowing the macro-system to
overwrite built-in runes.  For example, `=` is built-in syntax, but it
can later be extended by creating a macro which shaddows that rune.

This patch takes advantage of that to move to a simpler, faster, and
more-easily-jet-matched representation.