cog: start changing the filter to a more proper sandbox.
In addition to filtering unsafe usage of `[%cog %stop ...]` and `[%cog
%reap ...]`, also give a way to filter requests based on a list of
request prefixes. These are turned into a more efficient at runtime
format and then every request which doesn't match one of these is
treated as if it's an invisible 0 request.

Also fixes %stop/%reap so that they unwrap the sandbox function when
handing the data back for a cog they launched.