jelly: faster frag loader
This rewrites the fragment loader of Haskell's Jelly implementation.
The "fragments" are the actualy bit-encoded tree section of a jelly file.

The previous implementation worked against a lazy stream of booleans.
This instead uses pointer and bit-shift operations against an array
of Word64.

This halves Jelly load time, and reduces overall Sire load-from-cache
time by 25%.