plunder: remove machine names
This is part one in removing the complicated, multi-tenant plunder server
code to move to the "one machine per process model". It removes the concept
of MachineNames from the LMDB database, the Machine runtime code, and parts
of the ServerExe.hs code. But we're currently stubbing out several parts of
the ServerExe code with false machines to get it to compile before the
heavy overhaul of ServerExe in the next patch.