doc: fix an edge-case in the PLAN spec
The new PLAN spec failed to handle the edge-case when "running" a pin
that isn't a law.

The old spec handled this correctly, and the main prototype implementation
also handles this correctly, but the new pseudo-code spec missed this case.

For example, this input was handled incorrectly:

    (<(2 0)> 3 4)

The X() function correctly chooses 2 as the head, but runs that against
the environment as-is, but that environment has only two parameters and
the 2 primop requires 3.

This ended up being fixable pretty easily by just pre-processing the
environment before doing substitution with X()

    S(o:(f x y))       = (S((f x)) y)
    S(o:(<{n a b}> y)) = o
    S(o:(<f> x))       = S((f x))
    S(o)               = o

This transforms the closure into the following:

    (2 0 3 4)

And then runs X() substitution on that.

It's safe to mutate the closure in this way, because this happens during
evaluation, which mutates the closure regardless.