rts: standardize array-index operators
Previously, (^) was exponent in some situations and `indexSmallArray`
in others.  (!) was vector index, and no operator was defined for
`indexArray`

This changes `indexSmallArray` to (.!) and `indexArray` to (!).  This is
uglier for now, but is consistent.

Subsequent patches switch a lot of code from `Vector` to `Array`, which
will make resolve most of the ugliness introduced here.