rts: rows are arrays, not vectors.
Note that V.fromArray is basically free, and V.toArray is free if the
vector has not been sliced.

This removes three words and an indirection from every row, and paves
the way to directly using the GHC primitive "Array#" for Rows.

After this change, the `take` and `drop` jets are O(n), not O(1).

Also, if we retain a small slice of a large array, that no longer retains
the unreferenced memory.