sire: avoid iterative tabIns while parsing json
Instead of using tabIns to accumulate a tab while parsing objects,
we accumulate a list of pairs and then used that to create a table at
the end.

The change in performance is well below the (substantial) noise floor,
but this approach is O(n*log(n)) instead of O(n^2).