hitch: in intersection, operate on a lazy tree.
Intersection requires a tree without any hitchhikers. So we must force them
down before we operate. But previously, we were using the normal hitchhiker
nodes, which pin their children, and pinning immediately forces evaluation.

So make a structure of thunks and work with that instead so we don't
strictly make nodes that will never be evaluated.