rex: permit line-strings in closed forms
This change permits inputs such as:

    b#""" Hello.  How are you doing?

This only works in a handful of situations, because in cases like:

    ("foo", """bar)

The `""bar)` line-string consumes the entire input and there is no way
to write the final end-parenthesis.

However, that edge case adds no complexity, and the syntax is nice in
some cases.