rts: `plunder repl` lets you talk to seed files
This adds a new `repl` subcommand to the plunder exe, which can be used
like so:

    $ plunder repl echo.seed
    $ plunder repl rex.seed
    $ plunder repl sire.seed

This feeds passes bytes read from stdin into a simple PLAN machine.

    type REPL = Bar -> (Bar, Repl)

And feeds the output back to stdout.

The PLAN machine is responsible for splitting lines, etc.  Like with the
`read()` syscall, we indicate EOF by passing an empty bytestring.