When a cog has no more requests, it should be visible to %reap.
A cog in the state where it has zero requests and has exited cleanly
should be detectable by the cog lifecycle machinery exposed to cogs: it's
in the interest of any kernel to be able to detect programs that have
cleanly shut down.

This is a separate concept from whether the interpreter can tick a cog:
the interpreter might just not understand a request and that should
shut down the interpreter. But that shutdown shouldn't be visible from
inside the machine: you don't leak information about the interpreter to
the inside.