demo: webapp that can launch and halt cogs
This adds a new `sh/runcog-demo` script, which when run starts
a machine with a single cog and uploads the elm interface from
elm/runcog-demo to it.

This demo is currently able to click a button in the browser and
launch a predetermined demo cog on the machine. You can start as
many cogs as you want, though the cogs are all the same compiled
in test cog. (Once seed-in-sire is merged, this will accept a
seed file upload.)

Once launched, you can click Stop next to the list of actions for
that running cog.