seed: bad seed_save() crashes in haskell
If seed_save() writes more/less bytes than the pre-computed
buffer size, that's an internal variant violation and we need
want to crash.

This change makes this case throw an exception in Haskell
which prints out the value.  Previously, this just directly
killed the process from C with a cryptic error message, which
made debugging very difficult.

Hopefully, this case never actually happens again, but let's
keep these here until we're sure that the code is totally
correct.