sire: update cache even on error
Previously, if we encountered an error, the Sire build cache
was never updated.  This was true even if several modules were
loaded successfully before the offending one.

With this change, if any modules were loaded successfully,
the Sire cache is always updated to include those modules.

Implementation
--------------

Previously, the PLAN value representing the cache was threaded
through execution, and the final value was written to a file
(if it contained any changes).

Now, the cache value is stored in an `IORef` and that ref is
mutated whenever a module load successfully.  The whole process
is wrapped in a `bracket` so that the final "save if the cache
has changed" logic runs even if there is an error.