sire: CLI gains support for preloads
Previously it was difficult to test that all sire files load
correctly.  Because each sire file has it's own load path,
and there are multiple "forks" in the load-sequence, we needed
to separately test each end-node.

This patch introduces a concept of "preloads" to the Sire CLI, so if
you say:

    sire sire/foo.sire sire/bar.sire sire/zaz.sire

Then "foo" and "bar" are pre-loaded and added to the cache,
and then "zaz" is loaded and we enter a REPL at the end of the
"zaz" module.

The whole process crashes when we hit the first error, and the whole
process runs within a single cache context.

So now, running `sire/*.sire` will load all modules and crash
on any errors, and that will happen very efficiently because everything
is cached.

This patch also introduces this shorthand:

    `sire prelude` ==> `sire sire/prelude.sire`

And this also updates "sh/sire-load-all" to preload `*.sire`.