sire: rename the `` macro to '
Before this patch, the {`} macro is "quasi-quotation", but the "raw
quotation" macro is {``}.  This is ugly, and doesn't match the lisp
convention of using {'} for this.

With the recent changes to Rex, {'} is now a runic character, so we can
switch "raw quotation" to the prettier and more conventional rune.