exo: fix bug in `shiftR`
There was a bug in `shiftR` in `Exo`, our custom wrapper
library around GMP bignats.

Before this patch, when attempting to shift right by a multiple
of 64, this would perform a word slice (just ignoring `n`
least-significant words) and then running `mpn_rshift` to
shift by zero bits.

This was undefined behavior.  This patch changes the behavior in
that case to simply copy the remaining words with `mpn_copyi`.