plan: remove zero-argument laws
Previously, PLAN supported zero-argument laws where the 0=self reference
referred to the result of the computation (in the same way that a
self-recursive binding can refer to the result of the expression).

This was annoying to implement, especially in a native runtime system.
Instead of creating a new function and generating code for that function,
it just runs the law body with a cycle.

But even more importantly, this was not useful and people
pretty universally found the concept highly confusing.

This patch changes the behavior of zero-argument laws to simply
crash, updating the spec, all of the demo implementations, and the
Haskell runtime.

The following changes were also required:

-   The `_Force` jet (and the `_Trk` jet by proxy) previously relied on
    this behavior, and was updated.

-   The Sire compiler previously compiled top-level expressions into
    zero-argument laws, and now instead compiles them to one-argument
    law and passes in a dummy argument.

-   Once piece of code in `pq.sire` and one test took advantage of
    the behavior.  These were updated to use LETREC instead.