sire: rewrite inlining routine
This reworks the Sire inline routine to be much less code and
to handle more cases.

The behavior is not identical to the old code.  In particular,
it floats let bindings outside of applications in some cases.
However, everything still works and none of the jet hashes
change.

There is work in progress on a document that specifies the inlining
behavior in extreme detail, but that's not yet ready to merge, so the
behavior remains undocumented in this patch.