sire: some new utilities on lists
`listFoldl1` folds over a list, using the first element as the starting
state and crashing on empty lists.

`listRep` creates a list by repeating a single elemenet.

`listIdxCps` takes continuations for `not_found` and `found`
and `listIdxMb`, `listIdxOr` and `listIdx` are implemented as
specializations of that.

I'm not yet sure if this the right way to organize things.