sire: rex-in-sire
This patch introduces a full implementation of Rex in Sire,
along with tests that show that it has identical behavior to
the Haskell impl for all of our existing Sire files.

The test works by compiling a "Rex REPL" into a seed file,
and then parsing/printing each sire file with `plunder repl`.