sire: rewrote the backend of the Sire compiler
This is a rewrite of the Sire compiler backend.

The new approach moves all of the {@}s to the top-level of
each law and reorders them based on a traversal of the code
(ignoring the order of {@}s in the input).

This for example:

    ? (f a b)
    @ x (a b)
    | if even-a
        @ y (b a)
        | f (y y) x
    | else
        x

Becomes this:

    ? (f a b)
    @ y (b a)
    @ x (a b)
    | if even-a
        | f (y y) x
    | else
        x

This is an invasive change and breaks many jet hashes, however:

-   Is is much less complicated and requires much less code.

-   The resulting code works on the current version of PLAN, but will
    also work without modification with the pending changes to PLAN.