Reorder cog effect handling to fix hardware
If a newly spun cog immediately starts with a call to hardware, the
hardware would see that call before it gets notified that a new cog
had startted spinning. This is now fixed by changing the order we do
all the work in runResponse.

This also just cleans up and documents the entire block.