rts: fix seed tests
This patch fixes a bug in the way that the "number of bytes written" was
calculated in `seed_save`.  The bug was causing the seed tests to fail.

In the case where the value being saved was a single leaf, and that leaf
took less than a word to write, there was a mismatch in the way that
`seed_save` and `seed_size` calculated the result.

`seed_size` rounded up to the nearest multiple of 8 (since the result is
always an array of words).  `seed_save` just returns the actual number
of bytes written.

Since the buffer is pre-filled with zeros before saving, there was no
actual mismatch in the semantics.  The fix was to just have `seed_save`
also round up in this case.