rts: generalize seed API
Previously, all of the seed APIs worked in terms of pins, even when there
was no need for that.  That resulted in almost all of the call-sites
wrapping things with `mkPin` and unwrapping things with `pin.item`.

This patch changes those interfaces to work with arbitrary nouns.
There is no behavioral change, since the actual implementations were
simply extracting the the pin item and working with that.