boot: wutE learns about pins and marks
`wutE` is a utility for creating generating Sire expressions from macros.

Before there were two variants of this, `wutE` was unpinned (the `#?`
rune), and `wutWutE` was pinned (the `#??` rune).

With this patch, these are combined, and the combination is extended to
also support "always-inline" annotations.  So, now there is a flag for
"pinned" and a flag for "always-inline".