rts: terse constructor names for the Sire datatype
This change is mostly inspired by having a really hard time wading through
debug output with the longer names, and then having a much easier time
once switching to the short names.

This hurts legibility slightly, but this datatype is so central
to Sire backend, that any serious attempt to understand that code
will requires you to become pretty familiar with this datatype.