sh: note about sh/ci ignoring sire errors
Sire throws an exception when it errors out, but this exception
is caught and used for printing, so the executable actually does
not exit with a non-zero return code.

This violates the assumptions of the sh/ci script, and this leads
to erroring sire code not causing the CI script to fail.