rts: fix a ref-counting bug in the sire backend
The `stats` routine in `Sire.Backend` was not correctly
traversing the bindings graph.

The problem was that we didn't properly track the set of
"processed nodes", only treating them as "seen" once fully
processed (instead of as soon as we *begin* processing them).

This caused double-counting of references in some cases (which
caused some single-use bindings to not be eliminated), and it
also caused mutually recursive let-bindings to infinite loop in
some cases.

This patch resolves the issue by introducing a "Seen" set to the
state, and updating that immediately before we finish processing
the node. This also updates the jet hashes to reflect the change.